Fix : summary 에서 다시 이관작업 ...
